Move util functions out of __init__
